requirements.txt
Django==4.1.1
django-modeltranslation==0.18.4